operator: Merge gpu, gpu_openacc, and gpu_openmp
